DDF-3849 Refactored CsvQueryResponseTransformer into utility jar
