+++ catalog/catalog-app/src/main/resources/features.xml
@@ -246,8 +246,9 @@
     </feature>
 
     <feature name="catalog-transformer-csv" version="${project.version}"
-             description="CSV Query Response Transformer converts query results to CSV format.">
+             description="CSV Query Response Transformer and Common Library.">
         <bundle>mvn:ddf.catalog.transformer/csv-queryresponse-transformer/${project.version}</bundle>
+        <bundle>mvn:ddf.catalog.transformer/catalog-transformer-csv-common/${project.version}</bundle>
     </feature>
 
     <feature name="catalog-transformer-geoformatter" version="${project.version}"
