+++ catalog/transformer/pom.xml
@@ -88,7 +88,6 @@
         </dependencies>
     </dependencyManagement>
     <modules>
-        <module>catalog-transformer-csv</module>
         <module>catalog-transformer-attribute</module>
         <module>catalog-transformer-service-xslt</module>
         <module>catalog-transformer-resource</module>
@@ -112,5 +111,7 @@
         <module>catalog-transformer-zip</module>
         <module>catalog-transformer-propertyjson-metacard</module>
         <module>catalog-transformer-preview</module>
+        <module>catalog-transformer-csv-common</module>
+        <module>catalog-transformer-csv-queryresponse</module>
     </modules>
 </project>
